requirements.txt
fastapi~=0.104.1
uvicorn~=0.24.0
langchain~=0.1.0
langchain-google-genai~=0.0.6
chromadb~=0.4.22
python-dotenv~=1.0.0
sqlalchemy~=2.0.23
google-generativeai~=0.3.1
beautifulsoup4~=4.12.2
requests~=2.31.0
python-multipart~=0.0.6
pydantic~=2.5.2
unstructured~=0.11.0
pdf2image~=1.16.3
pytesseract~=0.3.10
tabulate~=0.9.0
pandas~=2.1.4
openpyxl~=3.1.2
